Reorganized ionosphere state functions to make it easier to find where to add new functions.
